调整message结构
