+++ include/aquarius/response.hpp
@@ -1,13 +1,13 @@
 ﻿#pragma once
-#include "message.hpp"
-#include "header.hpp"
+#include "detail\message.hpp"
+#include "detail\header.hpp"
 
 namespace aquarius
 {
 	namespace aqnet
 	{
 		template<class Body, std::size_t Number>
-		class tcp_response : public response<tcp_response_header, Body, Number>
+		class tcp_response : public response<Body, Number>
 		{
 		public:
 			virtual int parse_bytes(const void*, std::size_t)
@@ -22,7 +22,7 @@ namespace aquarius
 		};
 
 		template<class Body, std::size_t Number>
-		class http_response : public response<http_response_header, Body, Number>
+		class http_response : public response<Body, Number>
 		{
 
 		};
